booking_service/requirements.txt
pyodbc
redis
PyJWT
pika